# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API preliminar del Marketplace",
        "description": "Documentación generada con L5-Swagger para la API del Marketplace Orientado a la ciudad de Chivilcoy.",
        "contact": {
            "name": "Santiago Castellano",
            "email": "scastellano10@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/businesses": {
            "get": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Listar negocios del usuario autenticado",
                "description": "Devuelve todos los negocios asociados al usuario autenticado, incluyendo sus categorías e imágenes.",
                "operationId": "b47cb8f8dbb2224c2cc513bab2da85d8",
                "responses": {
                    "200": {
                        "description": "Lista de negocios del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Business"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Crear un nuevo negocio",
                "description": "Crea un nuevo negocio asociado al usuario autenticado. Valida el límite de negocios según la suscripción del usuario.",
                "operationId": "09636680657c34a3c34227aa276605d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Panadería San Jorge"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Panadería artesanal con más de 20 años de experiencia"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Falsa 123"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -34.6037,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -58.3816,
                                        "nullable": true
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Negocio creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para crear más negocios según su suscripción"
                    },
                    "422": {
                        "description": "Error de validación de los datos enviados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{business}": {
            "get": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Mostrar un negocio específico",
                "description": "Devuelve la información detallada de un negocio por su ID. Si el usuario autenticado es el dueño, también incluye los productos del negocio.",
                "operationId": "feb0cf913a78b75503617b23b3a7d154",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del negocio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Actualizar un negocio",
                "description": "Actualiza la información de un negocio. Solo el dueño del negocio puede actualizarlo.",
                "operationId": "cd3ec13e27b92b89bb83a6dc56aafd20",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Panadería San Jorge (Actualizado)"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Panadería artesanal con más de 25 años de experiencia"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Falsa 456"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -34.6037,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -58.3816,
                                        "nullable": true
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Negocio actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para actualizar este negocio"
                    },
                    "422": {
                        "description": "Error de validación de los datos enviados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Eliminar un negocio",
                "description": "Elimina un negocio existente. Solo el dueño del negocio puede eliminarlo.",
                "operationId": "1ba49b80ee03bbaf297a4bd08e88982f",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Negocio eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para eliminar este negocio"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{business}/categories/{category}": {
            "delete": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Eliminar una categoría de un negocio",
                "description": "Elimina una categoría específica de un negocio. Solo el dueño del negocio puede realizar esta acción.",
                "operationId": "cc9d36e893b6b97a015c6eca4f79d12d",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría eliminada del negocio correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoría eliminada del negocio correctamente"
                                        },
                                        "business": {
                                            "$ref": "#/components/schemas/Business"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para eliminar categorías de este negocio"
                    },
                    "404": {
                        "description": "El negocio no tiene asignada esta categoría"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{business}/categories": {
            "put": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Actualizar las categorías de un negocio",
                "description": "Actualiza todas las categorías asociadas a un negocio. Solo el dueño del negocio puede realizar esta acción.",
                "operationId": "72531511fb2d2c1ac929a6fa06b7a9cc",
                "parameters": [
                    {
                        "name": "business",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categorías actualizadas correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categorías actualizadas correctamente"
                                        },
                                        "business": {
                                            "$ref": "#/components/schemas/Business"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para actualizar categorías de este negocio"
                    },
                    "422": {
                        "description": "Error de validación de los datos enviados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/search": {
            "get": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Buscar negocios",
                "description": "Busca negocios por nombre, categoría o ubicación.",
                "operationId": "69c06fceafc425283c96de9a39ce3856",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre del negocio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "ID de la categoría",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitud para búsqueda por ubicación",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitud para búsqueda por ubicación",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radio de búsqueda en metros",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de negocios encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Business"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/businesses/nearby": {
            "get": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Negocios cercanos",
                "description": "Devuelve los negocios cercanos a una ubicación específica, dentro de un radio determinado.",
                "operationId": "b219f678dc17a7124cf73b3486990c4d",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitud",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitud",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radio de búsqueda en metros (por defecto: 10000)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de negocios cercanos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Business"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación de los parámetros de ubicación"
                    }
                }
            }
        },
...
        "/api/subscription": {
            "get": {
                "tags": [
                    "Suscripciones"
                ],
                "summary": "Obtener suscripción actual",
                "description": "Devuelve la suscripción actual del usuario autenticado. Si no tiene suscripción, crea una por defecto (plan 'free').",
                "operationId": "03430c7a696aa2c8bcd5cb67d41d9698",
                "responses": {
                    "200": {
                        "description": "Suscripción del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}